requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django-debug-toolba==3.6.0
django-phonenumber-field==7.0.0
phonenumbers==8.12.55
PyYAML==6.0
uritemplate==4.1.1
